requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0

# Search and text processing
rank-bm25>=0.2.2
jieba>=0.42.1

# UI
streamlit>=1.28.0
plotly>=5.18.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Diff and hashing
deepdiff>=6.7.0
xxhash>=3.4.0

# Optional: LLM integration (comment out if no API key)
# openai>=1.0.0
# anthropic>=0.18.0

# Visualization for PythonExecTool
matplotlib>=3.7.0
scipy>=1.11.0
